refactor(spec): extract lint-catalog and co-locate lint rule tests
Move pure LINT_CATALOG data into lint-catalog.ts (re-exported from lint.ts)
and split the long lint suite into rule characterization vs limits/wiring.
